Renamed Buffer middleware to RequestBodyBuffer
